refactor: TextInput 컴포넌트 분리
