+++ src/components/index.ts
@@ -2,6 +2,7 @@ export {default as Alert} from './Alert';
 export {default as Avatar} from './Avatar';
 export {default as Button} from './Button';
 export {default as Icon} from './Icon';
-export {default as Input} from './Input';
+export {default as FormInput} from './FormInput';
 export {default as SearchBar} from './SearchBar';
 export {default as Text} from './Text';
+export {default as TextInput} from './TextInput';
